+++ Baggage_Sorting_Unit/IAirplaneLoadingManagementt.java
@@ -1,4 +1,4 @@
-package BaggageSortingUnit;
+package Baggage_Sorting_Unit;
 
 public interface IAirplaneLoadingManagementt {
     void optimizeBalancing();
