+++ Baggage_Sorting_Unit/IBaggageSortingUnitRoboter.java
@@ -1,4 +1,4 @@
-package BaggageSortingUnit;
+package Baggage_Sorting_Unit;
 
 public interface IBaggageSortingUnitRoboter {
     void moveBaggageToDepot(ArrayList<Baggage> baggageList);
