+++ Baggage_Sorting_Unit/LoadingStrategy.java
@@ -1,4 +1,4 @@
-package BaggageSortingUnit;
+package Baggage_Sorting_Unit;
 
 import java.util.ArrayList;
